major update (see description)
Figured out what each class does. For TaxiClusters.java, wrote the code that takes in the CSV file and puts it into a List data structure. Completed GPScoord.java as well. Working on doing Cluster.java and trying to cluster points surround firstCoord together. NOTE: GPScoord cannot use String as that cannot be computed. It cannot be float because the accuracy is off.