COMPLETED (needs major refactoring, see description)
results are correct nowgit add . The problem was that I separated corePoint and surroundingCoords in Cluster.java instead of combining them into ONE LIST. The results are now correct. The CSV data has been checked with the sample output on text-compare.com. Needs code refactoring as attribures are currently public and commented code are everywhere.