IMPORTANTgit status (SEE DESCRIPTION)
This version REMOVES the P != Q in rangeQuery() and outputs the data DIFFERENTLY. It outputs a new CSV file PER CLUSTER CREATED as opposed to creating ONE LARGE CSV file that contains all clusters. The 700+ results is then sorted from largest to smallest, and the top 10 largest CSV file are then imported into Google Maps (https://www.google.ca/maps/about/mymaps/) to be plotted and compared to Robert Laganière's results (EPS = 0.0003, MIN_PTS = 5). The results appear to be IDENTICAL which means the P!=Q condition is different or wrong and does not apply to our project.