Key update (ignore last update and see description of this update)
Last update was not detailed enough, ignore it. Moving on. The plan is to user a single/multiple producer multiple consumer variation of the producer-consumer pattern. Multiple consumer is required, so single or multiple producer can be decided at a later commit. The link added is where it will teach this pattern. ADDITIONALLY Parameters are now passed by reference. Although only the seedSet in addToSeed() needs to be a pointer, I made everything a pointer to reduce the time and memory it takes to copy the large structs every time they are passed (especially the whole CSV file).