EFFICIENCY UPDATE; two major updates (see description)
Two changes that significantly boosted the performance. First: no longer using math.pow(num, 2) as it is slow in Go. Now using num*num. Second: No longer using the contains() function. Instead, removeDuplicateGPS() is used as a map is much faster to search than iterating through a slice. This person asked the EXACT SAME QUESTION as the problem encountered in this program. https://stackoverflow.com/questions/66643946/how-to-remove-duplicates-strings-or-int-from-slice-in-go