reduced time from 9 minutes to 8 seconds...
commented out the removeDUplicates() method. It's not needed because the 'if label(Q) != undefined then continue' condition in the inner for loop of DBSCAN (see Wikipedia's pseudocode for reference). That condition prevents duplicates, so checking for duplicatvents duplicates, so checking for duplicates or using \!contains() while adding seedNeighbours to seedSet is completely redundant.